LG-2505: Refresh radio box group (#1462)
* update for darkmode

* changeset

* fix box-shadow

* use rgba instead of transparent

* update test

* add to website

* feedback
