+++ .changeset/nice-bees-hope.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/radio-box-group': major
+---
+
+Updates `RadioBoxGroup` for dark mode brand refresh. Exports `RadioBoxProps` and `RadioBoxGroupProps`.
