+++ packages/radio-box-group/src/RadioBoxGroup.story.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Meta } from '@storybook/react';
 import Tooltip from '@leafygreen-ui/tooltip';
-import LeafyGreenProvider from '@leafygreen-ui/leafygreen-provider';
 import { RadioBoxGroup, RadioBox } from '.';
-import { RadioBoxGroupProps } from './RadioBoxGroup';
+import { RadioBoxGroupProps } from './types';
+import { storybookArgTypes } from '@leafygreen-ui/lib';
 
 export default {
   title: 'Components/RadioBoxGroup',
@@ -13,22 +13,26 @@ export default {
     onChange: { control: false },
     name: { control: 'text' },
     value: { control: 'text' },
+    darkMode: storybookArgTypes.darkMode,
+  },
+  parameters: {
+    controls: {
+      exclude: ['className', 'children', 'onChange', 'name', 'value'],
+    },
   },
 } as Meta<typeof RadioBoxGroup>;
 
 export const Uncontrolled = (args: RadioBoxGroupProps) => (
-  <LeafyGreenProvider>
-    <RadioBoxGroup name="radio-box-group-default" {...args}>
-      <RadioBox value="1">Option One</RadioBox>
-      <RadioBox value="2">Option Two</RadioBox>
-      <RadioBox default value="3">
-        Option Three
-      </RadioBox>
-      <RadioBox disabled={true} value="option-4">
-        Disabled Option
-      </RadioBox>
-    </RadioBoxGroup>
-  </LeafyGreenProvider>
+  <RadioBoxGroup name="radio-box-group-default" {...args}>
+    <RadioBox value="1">Option One</RadioBox>
+    <RadioBox value="2">Option Two</RadioBox>
+    <RadioBox default value="3">
+      Option Three
+    </RadioBox>
+    <RadioBox disabled={true} value="option-4">
+      Disabled Option
+    </RadioBox>
+  </RadioBoxGroup>
 );
 
 export const Controlled = (args: RadioBoxGroupProps) => {
@@ -43,34 +47,35 @@ export const Controlled = (args: RadioBoxGroupProps) => {
   );
 };
 
-export const TooltipTest = () => {
+export const TooltipTest = ({ darkMode }: RadioBoxGroupProps) => {
   return (
-    <LeafyGreenProvider>
-      <RadioBoxGroup>
-        <Tooltip
-          justify="middle"
-          trigger={<RadioBox value="one">One</RadioBox>}
-        >
-          This one is pretty cool.
-        </Tooltip>
-        <Tooltip
-          justify="middle"
-          trigger={<RadioBox value="two">Two</RadioBox>}
-        >
-          This does a thing.
-        </Tooltip>
+    <RadioBoxGroup darkMode={darkMode}>
+      <Tooltip
+        darkMode={darkMode}
+        justify="middle"
+        trigger={<RadioBox value="one">One</RadioBox>}
+      >
+        This one is pretty cool.
+      </Tooltip>
+      <Tooltip
+        darkMode={darkMode}
+        justify="middle"
+        trigger={<RadioBox value="two">Two</RadioBox>}
+      >
+        This does a thing.
+      </Tooltip>
 
-        <Tooltip
-          justify="middle"
-          trigger={
-            <RadioBox value="three" disabled>
-              Three
-            </RadioBox>
-          }
-        >
-          This is disabled for a reason.
-        </Tooltip>
-      </RadioBoxGroup>
-    </LeafyGreenProvider>
+      <Tooltip
+        darkMode={darkMode}
+        justify="middle"
+        trigger={
+          <RadioBox value="three" disabled>
+            Three
+          </RadioBox>
+        }
+      >
+        This is disabled for a reason.
+      </Tooltip>
+    </RadioBoxGroup>
   );
 };
