+++ packages/radio-box-group/src/RadioBoxGroup.styles.ts
@@ -1,0 +1,5 @@
+import { css } from '@leafygreen-ui/emotion';
+
+export const baseGroupStyle = css`
+  display: flex;
+`;
