--- packages/radio-box-group/src/Size.ts
@@ -1,9 +1,0 @@
-const Size = {
-  Default: 'default',
-  Compact: 'compact',
-  Full: 'full',
-} as const;
-
-type Size = typeof Size[keyof typeof Size];
-
-export default Size;
