+++ packages/radio-box-group/src/context.ts
@@ -1,11 +1,12 @@
 import { createContext, ChangeEventHandler, useContext } from 'react';
-import Size from './Size';
+import { Size } from './types';
 
 export interface RadioBoxGroupContext {
   value: string | number | undefined;
   onChange: ChangeEventHandler<HTMLInputElement>;
   size: Size | undefined;
   name: string | undefined;
+  darkMode: boolean;
 }
 
 const context = createContext<RadioBoxGroupContext | null>(null);
