+++ packages/radio-box-group/src/index.ts
@@ -1,3 +1,4 @@
 export { RadioBox } from './RadioBox';
 export { RadioBoxGroup } from './RadioBoxGroup';
-export { default as Size } from './Size';
+export type { RadioBoxProps, RadioBoxGroupProps } from './types';
+export { Size } from './types';
