+++ packages/radio-box-group/tsconfig.json
@@ -17,9 +17,6 @@
       "path": "../hooks"
     },
     {
-      "path": "../lib"
-    },
-    {
       "path": "../palette"
     },
     {
